Sample column specifications for reading CPS data

Because the CPS is a fixed-width file that changes data locations (and variable names) across years, to correctly read the data you have to specify which start/end positions correspond to which column names in each year. This is one such specification. To add extra data or change column names, see the Vignette.

Usage

cps_cols

Format

A data frame with 253 rows and 8 columns:

year

year

cps_name

original column name as given by the CPS

new_name

a new name, which tries to describe the variable and join sensibly across multiple years

start_pos

which character of a line the variable starts with

end_pos

which character of a line the variable ends with

col_type

whether the column is character, numeric, or a factor

description

the question text/description from the CPS

notes

any notes for question administration or analysis