Sample factor specifications for reading CPS data

Because the CPS changes factor levels across years, to correctly read the data you have to specify which numeric codes correspond to which character values in each year. This is one such specification. To add extra data, see the Vignette.

Usage

cps_factors

Format

A data frame with 2,475 rows and 5 columns:

year

year

cps_name

original column name as given by the CPS

new_name

a new name, which tries to describe the variable and join sensibly across multiple years

code

the numeric code contained in the raw CPS data

value

the character value corresponding to each numeric code

Details

These match the exact specifications from the CPS, including NA codes and any typos that occur (e.g., "Hipsanic" is common in older years).