recode the voting variable for turnout calculations

When the CPS calculates voter turnout, they consider the values "Don't know", "Refused", and "No response" to be non-voters, that is they lump these in with "No". With increased levels of survey non-response in recent years, this has caused turnout estimates to artificially deflate when compared to measures of voter turnout from state election offices. This function adds two recodes of the original voting variable, one which applies the CPS recoding where multiple categories map to "No", and one which follows the guidelines from Hur & Achen (2013) of setting these categories to NA. See the Vignette for more information on this process.

Usage

cps_recode_vote(
  data,
  vote_col = "VRS_VOTE",
  items = c("DON'T KNOW", "REFUSED", "NO RESPONSE")
)

Arguments

data

the input data set

vote_col

which column contains the voting variable

items

which items should be "No" in the CPS coding and NA in the Hur & Achen coding

Value

data with two columns attached, cps_turnout and hurachen_turnout, voting variables recoded according to the process above

Examples

cps_recode_vote(cps_refactor(cps_label(cps_2016_10k)))
#> # A tibble: 10,000 × 20
#>    FILE          YEAR STATE   AGE SEX   EDUCATION RACE  HISPANIC WEIGHT VRS_VOTE
#>    <fct>        <int> <fct> <int> <fct> <fct>     <fct> <fct>     <dbl> <fct>   
#>  1 cps_nov2016…  2016 AL       69 FEMA… HIGH SCH… WHITE NON-HIS…  1328. YES     
#>  2 cps_nov2016…  2016 AL       35 MALE  BACHELOR… WHITE NON-HIS…  1793. YES     
#>  3 cps_nov2016…  2016 AL       54 FEMA… HIGH SCH… WHITE NON-HIS…  1757. NO RESP…
#>  4 cps_nov2016…  2016 AL       47 MALE  HIGH SCH… WHITE NON-HIS…  1628. NO      
#>  5 cps_nov2016…  2016 AL       60 FEMA… SOME COL… WHITE NON-HIS…  1396. NO RESP…
#>  6 cps_nov2016…  2016 AL       12 FEMA… NA        WHITE NON-HIS…  1917. NA      
#>  7 cps_nov2016…  2016 AL       65 MALE  HIGH SCH… WHITE NON-HIS…  1732. NO      
#>  8 cps_nov2016…  2016 AL       43 MALE  SOME COL… WHITE NON-HIS…  2042. YES     
#>  9 cps_nov2016…  2016 AL       46 MALE  SOME COL… WHITE HISPANIC  2068. YES     
#> 10 cps_nov2016…  2016 AL       47 MALE  HIGH SCH… WHITE NON-HIS…  1694. NO      
#> # ℹ 9,990 more rows
#> # ℹ 10 more variables: VRS_REG <fct>, VRS_REG_WHYNOT <fct>,
#> #   VRS_VOTE_WHYNOT <fct>, VRS_VOTEMODE_2004toPRESENT <fct>,
#> #   VRS_VOTEWHEN_2004toPRESENT <fct>, VRS_REG_METHOD <fct>,
#> #   VRS_RESIDENCE <fct>, VRS_VOTEMETHOD_CON <fct>, cps_turnout <fct>,
#> #   hurachen_turnout <fct>