Calculations to reweight properly for voter turnout

While the U.S. Census Bureau provides one weight with the CPS, a modified weight is needed to properly calculate voter turnout. This data set provides those calculations, according to Hur and Achen (2013). The comparison data comes from Dr. Michael McDonald's estimates of voter turnout among the voting-eligible population (VEP). It can be joined with CPS data to calculate the new weights needed for analysis, using the function cps_reweight_turnout.

Usage

cps_reweight

Format

A tibble with 1,428 rows and 6 columns:

YEAR

year

STATE

state

response

indicator of turnout in recent election

vep_turnout

proportion of turnout indicator, calculated by McDonald

cps_turnout

proportion of turnout indicator, calculated by CPS

reweight

the factor by which to scale original CPS weights

Source

Turnout data from http://www.electproject.org/home/voter-turnout/voter-turnout-data